refactor: GameResultItem 컴포넌트 분리 #50
